+++ src/app/home/home.component.css
@@ -1,0 +1,26 @@
+.app-icon {
+    width: 40px;
+    margin-right: 10px;
+  }
+  
+  .header-fill {
+    flex: 1;
+  }
+  
+  .battery-level-indicator mat-icon {
+    margin-right: -4px;
+    vertical-align: text-bottom;
+  }
+  
+  .battery-level-indicator {
+    padding-right: 12px;
+  }
+  
+  mat-card {
+    margin-bottom: 10px;
+  }
+  
+  .connection-options {
+    margin-top: 16px;
+  }
+
